% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/env.R
\name{env}
\alias{clean_datacamp_env}
\alias{clean_everything}
\alias{clean_global_env}
\alias{clean_sct_env}
\alias{clean_solution_env}
\alias{env}
\alias{get_datacamp_env}
\alias{get_sct_env}
\alias{get_solution_env}
\title{DataCamp utility functions (environments)}
\usage{
get_datacamp_env()

get_solution_env()

get_sct_env()

clean_global_env()

clean_datacamp_env()

clean_sct_env()

clean_solution_env()

clean_everything()
}
\description{
Utility functions to retrieve relevant information for testing interactive
exercises on \url{www.DataCamp.com}. Environments.
}
\keyword{internal}

