% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exist.R
\name{exist}
\alias{exist}
\alias{exists_sct_result}
\alias{exists_solution_ds}
\alias{exists_solution_pd}
\alias{exists_student_ds}
\alias{exists_student_pd}
\title{DataCamp utility functions (do variables exist)}
\usage{
exists_sct_result(x)

exists_student_pd(x)

exists_solution_pd(x)

exists_student_ds(x)

exists_solution_ds(x)
}
\description{
Utility functions to retrieve relevant information for testing interactive
exercises on \url{www.DataCamp.com}. Do variables exist?
}
\keyword{internal}

