% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R
\name{get}
\alias{get}
\alias{get_chunk_number}
\alias{get_inline_number}
\alias{get_pec_code}
\alias{get_render_height}
\alias{get_render_width}
\alias{get_sct_code}
\alias{get_solution_code}
\alias{get_solution_ds}
\alias{get_solution_ds_part}
\alias{get_solution_output}
\alias{get_solution_pd}
\alias{get_student_code}
\alias{get_student_ds}
\alias{get_student_ds_part}
\alias{get_student_error}
\alias{get_student_output}
\alias{get_student_pd}
\alias{get_variable}
\title{DataCamp utility functions (getting variables)}
\usage{
get_variable(name, message = NULL, envir = get_datacamp_env())

get_pec_code()

get_student_code()

get_solution_code()

get_sct_code()

get_student_output()

get_student_error()

get_solution_output()

get_render_width()

get_render_height()

get_student_pd(x)

get_solution_pd(x)

get_student_ds(x)

get_solution_ds(x)

get_student_ds_part(x)

get_solution_ds_part(x)

get_chunk_number(x)

get_inline_number(x)
}
\description{
Utility functions to retrieve relevant information for testing interactive
exercises on \url{www.DataCamp.com}. Getting variables.
}
\keyword{internal}

