% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/global_vars.R
\name{global_vars}
\alias{exists_sct_result}
\alias{get_sct_result}
\alias{global_vars}
\alias{remove_sct_result}
\alias{set_sct_result}
\title{DataCamp utility functions (global env)}
\usage{
get_sct_result(x)

set_sct_result(x)

exists_sct_result(x)

remove_sct_result()
}
\arguments{
\item{x}{In set functions, this is the object you want to set.}
}
\description{
Utility functions to retrieve relevant information for testing interactive
exercises on \url{www.DataCamp.com}. Remove variables.
}

