% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remove.R
\name{remove}
\alias{remove}
\alias{remove_sct_result}
\alias{remove_solution_pd}
\alias{remove_student_error}
\alias{remove_student_output}
\alias{remove_student_pd}
\title{DataCamp utility functions (remove variables)}
\usage{
remove_student_output()

remove_student_error()

remove_sct_result()

remove_student_pd(suppress = FALSE)

remove_solution_pd(suppress = FALSE)
}
\description{
Utility functions to retrieve relevant information for testing interactive
exercises on \url{www.DataCamp.com}. Remove variables.
}
\keyword{internal}

