% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sct_vars.R
\name{sct_vars}
\alias{exists_solution_ds}
\alias{exists_solution_pd}
\alias{exists_student_ds}
\alias{exists_student_pd}
\alias{get_chunk_number}
\alias{get_inline_number}
\alias{get_solution_ds}
\alias{get_solution_ds_part}
\alias{get_solution_pd}
\alias{get_student_ds}
\alias{get_student_ds_part}
\alias{get_student_pd}
\alias{remove_solution_ds}
\alias{remove_solution_pd}
\alias{remove_student_ds}
\alias{remove_student_pd}
\alias{sct_vars}
\alias{set_chunk_number}
\alias{set_inline_number}
\alias{set_solution_ds}
\alias{set_solution_ds_part}
\alias{set_solution_pd}
\alias{set_student_ds}
\alias{set_student_ds_part}
\alias{set_student_pd}
\title{DataCamp utility functions (sct env)}
\usage{
get_student_pd(x)

set_student_pd(x)

exists_student_pd(x)

remove_student_pd()

get_solution_pd(x)

set_solution_pd(x)

exists_solution_pd(x)

remove_solution_pd()

get_student_ds(x)

set_student_ds(x)

exists_student_ds(x)

remove_student_ds()

get_solution_ds(x)

set_solution_ds(x)

exists_solution_ds(x)

remove_solution_ds()

get_student_ds_part(x)

set_student_ds_part(x)

get_solution_ds_part(x)

set_solution_ds_part(x)

get_chunk_number(x)

set_chunk_number(x)

get_inline_number(x)

set_inline_number(x)
}
\arguments{
\item{x}{In set functions, this is the object you want to set.}
}
\description{
Utility functions to retrieve relevant information for testing interactive
exercises on \url{www.DataCamp.com}. Setting variables.
}

