% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R, R/set.R
\name{get_sct_result}
\alias{get_sct_result}
\alias{set}
\alias{set_chunk_number}
\alias{set_inline_number}
\alias{set_pec_code}
\alias{set_sct_code}
\alias{set_sct_result}
\alias{set_solution_code}
\alias{set_solution_ds}
\alias{set_solution_ds_part}
\alias{set_solution_output}
\alias{set_solution_pd}
\alias{set_student_code}
\alias{set_student_ds}
\alias{set_student_ds_part}
\alias{set_student_error}
\alias{set_student_output}
\alias{set_student_pd}
\alias{set_variable}
\alias{set_variable_list}
\title{DataCamp utility functions (setting variables)}
\usage{
get_sct_result(x)

set_variable(name, var, envir = get_datacamp_env())

set_variable_list(lst)

set_pec_code(x)

set_student_code(x)

set_solution_code(x)

set_sct_code(x)

set_student_output(x)

set_student_error(x)

set_solution_output()

set_sct_result(x)

set_student_pd(x)

set_solution_pd(x)

set_student_ds(x)

set_solution_ds(x)

set_chunk_number(x)

set_inline_number(x)

set_student_ds_part(x)

set_solution_ds_part(x)
}
\description{
Utility functions to retrieve relevant information for testing interactive
exercises on \url{www.DataCamp.com}. Setting variables.
}
\keyword{internal}

