% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/uses_new_sct_system.R
\name{uses_new_sct_system}
\alias{uses_new_sct_system}
\title{Check if the current sct uses datacampSCT functions (FALSE) or testwhat functions (TRUE)}
\usage{
uses_new_sct_system()
}
\description{
Check if the current sct uses datacampSCT functions (FALSE) or testwhat functions (TRUE)
}

