% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{get_variable}
\alias{set_variable_list}
\alias{utils}
\title{DataCamp utility functions (helper function)}
\usage{
get_variable(name, envir, message = NULL)

set_variable_list(lst)
}
\description{
DataCamp utility functions (helper function)
}
\keyword{internal}

