% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datacamp_vars.R
\name{datacamp_vars}
\alias{datacamp_vars}
\alias{exists_pec_code}
\alias{exists_sct_code}
\alias{exists_solution_code}
\alias{exists_solution_output}
\alias{exists_student_code}
\alias{exists_student_error}
\alias{exists_student_output}
\alias{get_pec_code}
\alias{get_render_height}
\alias{get_render_width}
\alias{get_sct_code}
\alias{get_solution_code}
\alias{get_solution_output}
\alias{get_student_code}
\alias{get_student_error}
\alias{get_student_output}
\alias{remove_pec_code}
\alias{remove_sct_code}
\alias{remove_solution_code}
\alias{remove_solution_output}
\alias{remove_student_code}
\alias{remove_student_error}
\alias{remove_student_output}
\alias{set_pec_code}
\alias{set_sct_code}
\alias{set_solution_code}
\alias{set_solution_output}
\alias{set_student_code}
\alias{set_student_error}
\alias{set_student_output}
\title{DataCamp utility functions (datacamp env)}
\usage{
get_solution_code()

set_solution_code(x)

exists_solution_code(x)

remove_solution_code()

get_student_code()

set_student_code(x)

exists_student_code(x)

remove_student_code()

get_sct_code()

set_sct_code(x)

exists_sct_code()

remove_sct_code()

get_pec_code()

set_pec_code(x)

exists_pec_code()

remove_pec_code()

get_student_output()

set_student_output(x)

exists_student_output(x)

remove_student_output()

get_student_error()

set_student_error(x)

exists_student_error()

remove_student_error()

get_solution_output()

set_solution_output(x)

exists_solution_output()

remove_solution_output()

get_render_width()

get_render_height()
}
\arguments{
\item{x}{In set functions, this is the object you want to set.}
}
\description{
Utility functions to retrieve relevant information for testing interactive
exercises on \url{www.DataCamp.com}. Remove variables.
}

