% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/global_vars.R
\name{global_vars}
\alias{exists_sct_result}
\alias{exists_solution_output}
\alias{exists_student_error}
\alias{exists_student_output}
\alias{get_sct_result}
\alias{get_solution_output}
\alias{get_student_error}
\alias{get_student_output}
\alias{global_vars}
\alias{remove_sct_result}
\alias{remove_solution_output}
\alias{remove_student_error}
\alias{remove_student_output}
\alias{set_sct_result}
\alias{set_solution_output}
\alias{set_student_error}
\alias{set_student_output}
\title{DataCamp utility functions (global env)}
\usage{
get_sct_result(x)

set_sct_result(x)

exists_sct_result(x)

remove_sct_result()

get_student_output()

set_student_output(x)

exists_student_output(x)

remove_student_output()

get_student_error()

set_student_error(x)

exists_student_error()

remove_student_error()

get_solution_output()

set_solution_output(x)

exists_solution_output()

remove_solution_output()
}
\arguments{
\item{x}{In set functions, this is the object you want to set.}
}
\description{
Utility functions to retrieve relevant information for testing interactive
exercises on \url{www.DataCamp.com}. Remove variables.
}

