% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{uses_new_sct_system}
\alias{utils}
\title{DataCamp utility functions (other)}
\usage{
uses_new_sct_system()
}
\description{
Utility functions to retrieve relevant information for testing interactive
exercises on \url{www.DataCamp.com}. Other functionality.
}
\keyword{internal}

